Apex class: BatchOppClosedateAdd7days
Purpose: This Apex class is designed to update the Close Date and Subscription fee field on Opportunities that have not been closed and are not in the "Closed Won" stage.

Implements: This class implements the Database.Batchable interface.

Methods:

start(Database.BatchableContext BC) : This method retrieves a query locator for all Opportunities where the IsClosed field is false and the StageName is not "Closed Won".

execute(Database.BatchableContext BC, List<Opportunity> scope) : This method iterates through the list of Opportunities obtained from the start method, updates the CloseDate field to 7 days from today and Subscription_fee__c field to 25, and adds the updated Opportunities to a new list. The new list is then used to update the Opportunities in the database.

finish(Database.BatchableContext BC) : This method is an optional method that can be used to handle any post-processing that needs to be done after the execute method has completed. In this example, it is left empty as no post-processing is required.