calling future method from batch apex
programming example of calling future method from batch apex in salesforce ?

In this example, the BatchProcessor class implements the Database.Batchable interface and retrieves a list of Accounts using a query locator. In the execute method, the future method ProcessAccounts.processAccounts is called and passed the list of account IDs. The future method processAccounts can then perform the processing on the Accounts asynchronously, while the batch job continues to process the remaining records.